create xcode proj and add files
