+++ Swift-Essentials.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Workspace
+   version = "1.0">
+   <FileRef
+      location = "self:Swift-Essentials.xcodeproj">
+   </FileRef>
+</Workspace>
