+++ Swift-Essentials/Swift-Essentials.h
@@ -1,0 +1,19 @@
+//
+//  Swift-Essentials.h
+//  Swift-Essentials
+//
+//  Created by Aaron Dean Bikis on 7/25/17.
+//  Copyright © 2017 RadientTap. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for Swift-Essentials.
+FOUNDATION_EXPORT double Swift_EssentialsVersionNumber;
+
+//! Project version string for Swift-Essentials.
+FOUNDATION_EXPORT const unsigned char Swift_EssentialsVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <Swift_Essentials/PublicHeader.h>
+
+
